+++ examples/basic.js
@@ -18,6 +18,9 @@ const engine = new be.Engine({
     })
   })
 
+  commands.user('channel', 'leader')
+  commands.users('channel', 'members')
+
   commands.resources('channel', 'numbers', be.ResourceTypes.Number) // number list config
   commands.resources('channel', 'fruits', be.ResourceTypes.String) // string list config
