+++ lib/event-processors/command-event-processor.js
@@ -6,6 +6,7 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 var _ = require('underscore');
 var promises = require('../promises');
+var rt = require('../resource-types');
 
 var delegate = function delegate(prototype, to, name) {
   if (_.isArray(name)) {
@@ -18,9 +19,6 @@ var delegate = function delegate(prototype, to, name) {
     this[to][name].apply(this[to], arguments);
   };
 };
-
-var isPromise = promises.isPromise;
-var toPromise = promises.toPromise;
 
 var Commands = function () {
   function Commands(engine) {
@@ -141,18 +139,6 @@ delegate(Commands.prototype, 'controller', ['on', 'hears'])
     });
   }
 
-  function onShow(object, prop, type, entities) {
-    this.eventHandlers.push({
-      command: prop,
-      usage: prop + '\u3092\u8868\u793A\u3057\u307E\u3059',
-      handler: function handler(bot, message) {
-        return entities.getProp(message[object], prop).done(function (value) {
-          bot.reply(message, prop + ': ' + type.toString(value));
-        });
-      }
-    });
-  }
-
   Commands.prototype.resource = function (object, prop, type, entities) {
     var objects = object + 's';
     entities = entities || this.engine.storage[objects];
@@ -161,16 +147,25 @@ delegate(Commands.prototype, 'controller', ['on', 'hears'])
       command: 'set-' + prop + ' ' + type.format(),
       usage: prop + '\u3092\u8A2D\u5B9A\u3057\u307E\u3059',
       handler: function handler(bot, message) {
-        return toPromise(type.value(message)).then(function (value) {
+        return promises.toPromise(type.value(message, bot)).then(function (value) {
           return entities.saveProp(message[object], prop, value).done(function () {
-            bot.reply(message, prop + '\u3092\u8A2D\u5B9A\u3057\u307E\u3057\u305F: ' + type.toString(value));
+            bot.reply(message, prop + '\u3092\u8A2D\u5B9A\u3057\u307E\u3057\u305F: ' + type.toLabel(value));
           });
         });
       }
     });
 
     onClear.call(this, object, prop, type, entities);
-    onShow.call(this, object, prop, type, entities);
+
+    this.eventHandlers.push({
+      command: prop,
+      usage: prop + '\u3092\u8868\u793A\u3057\u307E\u3059',
+      handler: function handler(bot, message) {
+        return entities.getProp(message[object], prop).done(function (value) {
+          bot.reply(message, prop + ': ' + type.toLabel(value));
+        });
+      }
+    });
   };
 
   Commands.prototype.resources = function (object, prop, type, entities) {
@@ -181,23 +176,47 @@ delegate(Commands.prototype, 'controller', ['on', 'hears'])
       command: 'add-' + prop + ' ' + type.format(),
       usage: prop + '\u3092\u8FFD\u52A0\u3057\u307E\u3059',
       handler: function handler(bot, message) {
-        return toPromise(type.value(message)).then(function (value) {
+        return promises.toPromise(type.value(message, bot)).then(function (value) {
+          if (_.isUndefined(value)) {
+            return bot.reply(message, '有効な値が取れませんでした');
+          }
+
           var defaults = {};
           defaults[prop] = [];
           return entities.safeGet(message[object], defaults).then(function (entity) {
             entity[prop].push(value);
             return entities.save(entity);
           }).done(function () {
-            bot.reply(message, prop + '\u3092\u8FFD\u52A0\u3057\u307E\u3057\u305F: ' + type.toString(value));
+            bot.reply(message, prop + '\u3092\u8FFD\u52A0\u3057\u307E\u3057\u305F: ' + type.toLabel(value));
           });
         });
       }
     });
 
     onClear.call(this, object, prop, type, entities);
-    onShow.call(this, object, prop, type, entities);
+
+    this.eventHandlers.push({
+      command: prop,
+      usage: prop + '\u3092\u8868\u793A\u3057\u307E\u3059',
+      handler: function handler(bot, message) {
+        return entities.getProp(message[object], prop).done(function (value) {
+          bot.reply(message, prop + ':\n' + _.map(value, function (v) {
+            return type.toLabel(v);
+          }).join('\n'));
+        });
+      }
+    });
   };
 })();
+
+// aliases
+Commands.prototype.user = function (object, prop, entities) {
+  this.resource(object, prop, rt.User, entities);
+};
+
+Commands.prototype.users = function (object, prop, entities) {
+  this.resources(object, prop, rt.User, entities);
+};
 
 var CommandEventProcessor = function () {
   function CommandEventProcessor() {
