+++ lib/promises.js
@@ -1,13 +1,14 @@
 'use strict';
 
+var _ = require('underscore');
 var Q = require('q');
 
 function isPromise(o) {
-  return !!o.then;
+  return o.then;
 }
 
 function toPromise(o) {
-  if (isPromise(o)) return o;
+  if (!_.isUndefined(o) && isPromise(o)) return o;
 
   var deferred = Q.defer();
   deferred.resolve(o);
