+++ src/promises.es6
@@ -1,11 +1,12 @@
+const _ = require('underscore')
 const Q = require('q')
 
 function isPromise (o) {
-  return !!o.then
+  return o.then
 }
 
 function toPromise (o) {
-  if (isPromise(o)) return o
+  if (!_.isUndefined(o) && isPromise(o)) return o
 
   const deferred = Q.defer()
   deferred.resolve(o)
